Clarify CONTRIBUTING.md instructions on running e2e tests (#844)
Remove the reference to running `npm install -D @playwright/test`. It is
not necessary as that is listed in the package.json.
Correct `scripts` to `tests` in describing the end-to-end tests. It is
slightly more accurate.

These changes were supposed to be part of #832 but they were
accidentally omitted.

Fixes #838

<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot-pro/844"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>

Co-authored-by: Andrew M. Farrell <armorsmith42@gmail.com>