Run end-to-end tests on CircleCI (#843)
We will want end-to-end tests to run as part of our CI process.

This adds a new CircleCI job to do so using Playwright. It saves the
report and test results as artifacts. This includes Playwright traces
which CircleCI cannot currently display to the user.

Fixes #839.

versions of Python and Node.

To run tests against the dev server, Playwright needs a CircleCI image
with Python, Node and a browser. The build_python and build_javascript
steps use python-3.7.4 and node-14.17.5 respectively. The closest
versions we can get in a single CircleCI image for Playwright is
`python:3.7.11-browsers`, which has python-3.7.11 and node-14.17.1.

See:
https://circleci.com/developer/images/image/cimg/python#tagging-scheme.

One option for a piece of follow-on work is to update the versions of
the Docker images used for the jobs lint_sql, check_style,
build_javascript, and build_python.

<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot-pro/843"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>

---------

Co-authored-by: Andrew M. Farrell <armorsmith42@gmail.com>