Rename dev-tests folder to e2e (#860)
We typically use 'e2e' to refer to code relating to the e2e tests, so
this change renames the 'dev-tests' folder to 'e2e' for consistency.
'dev-tests' is a bit ambiguous, as unit tests are also a type of dev
test, but we keep those in other folders.
<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot-pro/860"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>