Parametrize spacing and sizing of on-screen keyboard (#1367)
Part of https://github.com/tiny-pilot/tinypilot/issues/1304, stacked on
https://github.com/tiny-pilot/tinypilot/pull/1366.

This is a preliminary, non-functional refactoring for a subsequent PR,
in which we want to adjust the sizing and spacing of the on-screen
keyboard. In order to make that process easier, I pulled out some shared
variables, and tried to make the underlying math more visible.

In theory, we could go even further with normalizing the parameters
between `<key>` and `<on-screen-keyboard>`, because they are specified
redundantly now. But then we’d have to pull out a standalone CSS file
that we import in both component, which would only contain a few CSS
variables. I wasn’t sure that would be worth it, so I thought a comment
would do just as well here.
<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot/1367"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>