Reduce sizing and spacing requirement of on-screen keyboard (#1368)
Part of https://github.com/tiny-pilot/tinypilot/issues/1304, stacked on
https://github.com/tiny-pilot/tinypilot/pull/1367.

Our goal is to eventually change the on-screen keyboard so that it
hovers over the lower part of the remote screen. Since the hovering
keyboard will cover up parts of the remote screen, we want to increase
the space efficiency in order to mitigate the covering effect.

In this PR, we slightly reduce the size of the keys, and the spacing
between the key blocks. As a result, the keyboard becomes a bit slimmer,
but it shouldn’t hinder the usability yet.
<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot/1368"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>