Add collapse button to on-screen keyboard (#1375)
Part of https://github.com/tiny-pilot/tinypilot/issues/1304; stacked on
https://github.com/tiny-pilot/tinypilot/pull/1374.

This PR adds a collapse button to the on-screen keyboard, so that the
keyboard can be stashed away more conveniently.


https://user-images.githubusercontent.com/83721279/234397126-d60ee922-eeec-42e9-83d4-7b68ad976df6.mov

Notes:

- We don’t (want to) have an equivalent shortcut for showing the
keyboard for now. This would probably have to live in the status bar,
and as discussed in [the
ticket](https://github.com/tiny-pilot/tinypilot/issues/1304), this is a
bit of a bigger topic in itself.
- Since the keyboard can be toggled from multiple places now, I had to
restructure the code around updating the view settings and menu bar
status. I thought it made most sense to leave the responsibility in
`app.js`, and couple it loosely via a callback (`onVisibilityChanged`).
<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot/1375"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>