Add an e2e test of the debug logs dialog (#859)
Resolves #840

Add a Playwright test to check that a browser can open the dialog under
`System > Logs` and copy its publicly-sharable logs URL to the
clipboard.

To check that the clipboard actually has the url after clicking the
copy-to-clipboard button, the test creates a new browser window with an
`<input>` element, simulates hitting `Control+V` to paste into the
`<input>`, and reads what it pasted. The test needs to create a new
window because the TinyPilot UI has event handlers which the capture and
discard the `Control+V` key press event.

I've created a separate test helper function for this. Normally I prefer
to wait much longer before splitting test code into a separate function.
I actually favor having more duplication in tests than I'd want in app
code, in order to make it easier to see what the test is doing.

However, because this is a workaround to a limitation of Playwright,
having a separate function seemed like a much clearer way to answer the
question "why am I creating an entirely new page and adding an element
to it?"

I don't (yet) check the contents of the logs because we need a reliable
way to get a real device to generate sensitive messages. (See
https://github.com/tiny-pilot/tinypilot-pro/issues/870)

<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot-pro/859"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>

---------

Co-authored-by: Andrew M. Farrell <armorsmith42@gmail.com>