Make background of on-screen keyboard slightly translucent (#1377)
Part of https://github.com/tiny-pilot/tinypilot/issues/1304.

Currently, the on-screen keyboard has a grey, static background ([see
demo video here,
e.g.](https://github.com/tiny-pilot/tinypilot/pull/1374)). This PR
suggests to make the background slightly translucent, to give the
keyboard drawer a more lightweight and hovering touch. That helps users
to perceive the UI as more 3-dimensional, as if the keyboard is floating
*above* the remote screen, rather than sliding *into* it.

That’s purely a visual effect, though – just as before, the hovering
keyboard of course covers part of the remote screen, and the blurry
background doesn’t and cannot fix this.


https://user-images.githubusercontent.com/83721279/234694772-c694cae2-5406-426b-93e9-5cc3e16b5c01.mov

It’s probably best to test this branch on the device in a real scenario,
otherwise it’s hard to get a feeling for it – e.g., on different remote
screen backgrounds, but also in different view port sizes.

This PR is not urgent, so it might also make sense to take a bit of time
to familiarize ourselves with the new hovering keyboard in general,
before we proceed to make more visual tweaks like this one.

Though I personally like this design, I’m not married to it, and think
the grey, static background works perfectly fine.

<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot/1377"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>