Add ESLint to the build (#1390)
This adds ESLint to the build. I'm surprised we didn't already have it!
I thought it was there all along.

This also makes a few changes in response to issues ESLint flagged. I
tried to keep the fixes limited to very low-risk changes. For anything
more substantial, I just suppressed the rule for that line.

A few are worth calling out:


https://github.com/tiny-pilot/tinypilot/blob/9beae18851c1bdde61fbe0de02020a6b134b573f/app/static/js/app.js#L75-L77

This just never worked. It was introduced in #326 and never referred to
anything. I think I just copy/pasted from the section where `resolve`
was defined and didn't realize it wasn't defined in this context.

I've replaced it with a `return;`.


https://github.com/tiny-pilot/tinypilot/blob/9beae18851c1bdde61fbe0de02020a6b134b573f/app/templates/custom-elements/remote-screen.html#L174-L182

This has been broken since #328 and nobody seems to have noticed, so
I've just removed this code.

<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot/1390"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>