Simplify placement of the usb-gadget service definition
As part of the War on Ansible, I'd like to refactor a little bit how we place the usb-gadget systemd file. We're unlikely to change the path of the TinyPilot privileged directory from the current /opt/tinypilot-privileged, so we should just hardcode it into the usb-gadget systemd file and save ourselves the extra Ansible logic of templating the file. This will make it slightly easier to later port the systemd file into debhelper like we currently do with tinypilot-updater.