Document dev workflows for testing on device (#1414)
Resolves https://github.com/tiny-pilot/tinypilot/issues/1073.

In our discussion in the [previous draft
PR](https://github.com/tiny-pilot/tinypilot/pull/1407) around dev
scripts and workflow documentation for remote testing, we realized that
it might be too tricky for us to provide easy-to-use dev scripts that
work in all environments and setups out of the box.

We could still add such dev scripts at a later point. For now, this PR
updates and extends our documentation. I think that’s also a good way to
share tips and workflows (although it’s prose), and to align ones own
workflows with best practices that others are using. I tried to cover
all topics that I could think of, and that appeared useful or important
to me. The structure is more or less loose, but I thought it makes sense
to group things into sections.

I also removed the section on SSH agent forwarding. This should be
obsolete since we abandoned the git-based update workflow.

If there is anything else that we can add, please let me know.
<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot/1414"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>