Remove 'exact' option from e2e test matchers
Our e2e code uses a mix of locators with the 'exact' option and locators that omit it.

The 'exact' option does a whole-string match and is case sensitive, whereas the default (or 'exact: false') does not.

In general, I prefer rigorous tests, but I don't think there's enough value in the 'exact' option to outweight the clutter it adds to the tests, so I move to drop it as the default and only use it in places where an exact match is especially important.