Add ARIA roles to menu bar elements (#1452)
This change adds four Accessible Rich Internet Applications (ARIA) roles
to our menu bar. This allows us to use clearer semantics in our
end-to-end tests and better present the UI to assistive technologies.

I largely based the role assignment on these examples:

*
https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/menuitem_role#examples
*
https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/group_role#examples

### `menubar`

>A menubar is a presentation of menu that usually remains visible and is
usually presented horizontally.


https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/menubar_role

I applied this to the top-level container for our menu.

I debated between `menubar` and `menu`, but `menubar` felt like the
closer fit.

### `menuitem`

>The `menuitem` role indicates the element is an option in a set of
choices contained by a `menu` or `menubar`.


https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/menuitem_role

I applied this to every `<a>` tag in our menu bar.

### `presentation`

>The `presentation` role and its synonym `none` remove an element's
implicit ARIA semantics from being exposed to the accessibility tree.

>The content of the element will still be available to assistive
technologies; it is only the semantics of the container — and in some
instance, required associated descendants — which will no longer expose
their mappings to the accessibility API.


https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/presentation_role

I applied this to every `<li>` element in our menu, as my understanding
is that we want to tell assistive tools that they're not *really* list
items.

### `group`

>The group role identifies a set of user interface objects that is not
intended to be included in a page summary or table of contents by
assistive technologies.


https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/group_role

I applied this to every `<ul>` in the menu that represented a group of
menu items.

---

I also changed our end-to-end tests to make more use of `getByRole`,
which is Playwright's recommended method for locating page elements:

>We recommend prioritizing role locators to locate elements, as it is
the closest way to how users and assistive technology perceive the page.

https://playwright.dev/docs/locators#locate-by-role

### Implementation note

There's pretty much a 1:1 mapping between our CSS classes and these new
ARIA roles. I started trying to replace the classes and changed the CSS
selectors to use the `role` attributes, but it got a little hairy, so I
decided to limit these changes to just adding the proper ARIA roles
without messing with CSS.

<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot/1452"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>