Revise installation section of README
Our install instructions assume an HDMI to USB dongle, so users with an HDMI to CSI chip try to install, they tend not to know that they need to perform extra steps.

This change revises the install section of the README to remove implementation details about what TinyPilot installs and make it more obvious how to find instructions for installing with an HDMI to CSI chip.