Remove remove_tc358743.yml tasks
This covers the case of the user switching their devices from from using the HDMI to CSI chip to the HDMI to USB dongle. I doubt that there are ever real-world scenarios for this, and if they are, it makes more sense to just do a clean install.

Removing this de-clutters our Ansible role by cutting out essentially dead code.