Get rid of uStreamer capture device Ansible vars
These are a lot of complexity to manage something pretty simple. Instead of keeping state in these confusing variables, we can just check if the string 'dtoverlay=tc358743' appears in /boot/config.txt. If it does, it's a TC358743-based system. Otherwise, it's an HDMI to USB dongle based system.