Remove remove_tc358743.yml tasks (#1553)
There was never really a need for this file. It covers the case of the
user switching their devices from from using the HDMI to CSI chip to the
HDMI to USB dongle. I doubt that there are ever real-world scenarios for
this. If there are, it makes more sense to just do a clean install, so
we don't need to offer a supported path between the two modes.

Removing this de-clutters our Ansible role by cutting out dead code.

<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot/1553"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>