Add menu item for opening dedicated window (#1589)
Part of https://github.com/tiny-pilot/tinypilot/issues/728.

This PR adds an item to the “View” menu, which is labelled “Dedicated
Window”, and which opens the remote screen in “standalone” view mode in
a popup window.


https://github.com/tiny-pilot/tinypilot/assets/83721279/f6bdbdf5-5387-4f6a-8401-9b33f92fb5cb

## Notes

- As discussed in [this issue
comment](https://github.com/tiny-pilot/tinypilot/issues/728#issuecomment-1692216456),
we decided to use a rather pragmatic and slightly unconventional
approach for hiding the remote screen from the main (parent) window, by
issuing an external redirect to a separate page. A “clean” and proper
teardown mechanism, which would allow us to stay on the main page, would
probably be rather complex to build, and also brittle to maintain.
- I’ve debated whether to use an `btn-action` (blue) or a `btn-success`
(green) for the “Refresh Page” button. I eventually went with the
former, since the action is optional, and the button is only a
convenience.
<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot/1589"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>