Replaces cryptography project test in e2e tests
Our e2e tests make sure that links work for a project that has a license hosted outside the device (at a web URL). Our current test uses cryptography as this project, but we're removing cryptography as a dependency in #1596 and #1604, so we need to pick another project to exercise in our e2e tests.