Bring back --allow-releaseinfo-change-suite
In #776, we added the --allow-releaseinfo-change-suite to avoid erroring out if the Raspbian apt repo changed suite names.

In #1604, we deleted the Reading package lists... command entirely, as it seemed unnecessary after deleting Ansible.

In #1623, we brought back the Reading package lists... command, but we left off the  flag we were previously using.