Reference style guide from CONTRIBUTING doc (#1737)
Related to (and stacked on)
https://github.com/tiny-pilot/tinypilot/pull/1736, I thought we could
also add a style guide reference to the CONTRIBUTING doc.

Note that we are not consistent with title case vs. non-title case for
headlines in that CONTRIBUTING doc. Most headlines don’t seem to use it,
so I’ve followed that for consistency.
<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot/1737"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>

---------

Co-authored-by: Jan Heuermann <jan@jotaen.net>