Dispatch DialogClosed event when clicking X button (#1828)
Related https://github.com/tiny-pilot/tinypilot/issues/1814. Stacked on
https://github.com/tiny-pilot/tinypilot/pull/1827.

This PR changes the control flow when clicking the `X` button in the
`<overlay-panel>`: instead of hiding the panel directly, it now
dispatches a `dialog-closed` event into the slotted element, to give the
slotted element a chance to hook into this event and react to it.

We need this in [a subsequent PR (the network status
dialog)](https://github.com/tiny-pilot/tinypilot/pull/1829), which needs
to stop its internal update loop once the dialog terminates. This also
must work if the user clicks the `X` button, not just when using the
`Close` one.

<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot/1828"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>

---------

Co-authored-by: Jan Heuermann <jan@jotaen.net>