Introduce .env file for storing configuration (#1874)
Related https://github.com/tiny-pilot/tinypilot-pro/issues/1446.

This PR introduces the ability to store configuration values in a `.env`
file, as explored and discussed in
https://github.com/tiny-pilot/tinypilot-pro/pull/1473. This allows us to
overwrite certain parameters to facilitate our development flow.

This is a significant change insofar as it introduces a new library. To
me, this would be justified, as I’d think that (a) `.env` files are
widely popular for managing configuration, and (b) our current
`.cfg`-based solution is lacking for our purposes. My thinking is to go
on and deprecate the `app_settings.cfg`, and eventually migrate the two
entries (`KEYBOARD_PATH` and `MOUSE_PATH`) to the `.env` file. If you’d
agree with this approach, I’d create a separate ticket for task.
<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot/1874"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>

---------

Co-authored-by: Jan Heuermann <jan@jotaen.net>