Fix wifi-dialog (#1921)
As [discovered
here](https://github.com/tiny-pilot/tinypilot-pro/issues/1618#issuecomment-3361292481),
we've broken the wifi-dialog via
https://github.com/tiny-pilot/tinypilot/pull/1902.

The issue is that we've previously changed the output of the
`/api/network/status` endpoint to return a list of network interfaces,
but we forgot to update the endpoint's usage in wifi-dialog.

This PR fixes the usage of the `/api/network/status` endpoint in
wifi-dialog.

<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot/1921"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>